% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{get_data}
\alias{get_data}
\title{A Data Fetching Function for Population Pyramids}
\usage{
get_data(country, year)
}
\arguments{
\item{country}{Enter ISO country code (e.g. 'IZ' for Iraq)}

\item{year}{Enter 4-digit year (e.g. 2012)}
}
\description{
This function pulls down census population data and puts it into a nice dataframe.
}
\examples{
get_data('IZ', 2012)
}
\keyword{age}
\keyword{census}
\keyword{dataframe}
\keyword{gender}
\keyword{population}
