% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pyramid.r
\name{plot_pyramid}
\alias{plot_pyramid}
\title{A plotting Function for Population Pyramids}
\usage{
plot_pyramid(country.frame, plot.title = "Population Pyramid")
}
\arguments{
\item{country.frame}{Enter dataframe generated from get_data()}

\item{plot.title}{(optional) title of plot}
}
\description{
This function pulls down census population data and puts it into a nice dataframe.
}
\examples{
plot_pyramid(iraq.frame, 'Iraq Population Pyramid')
}
\keyword{age}
\keyword{census}
\keyword{dataframe}
\keyword{gender}
\keyword{population}
